+++ kem/src/x25519/x25519.h
@@ -1,0 +1,39 @@
+/*
+ * Copyright (C) 2023, Stephan Mueller <smueller@chronox.de>
+ *
+ * License: see LICENSE file in root directory
+ *
+ * THIS SOFTWARE IS PROVIDED ``AS IS'' AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+ * OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE, ALL OF
+ * WHICH ARE HEREBY DISCLAIMED.  IN NO EVENT SHALL THE AUTHOR BE
+ * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+ * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT
+ * OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
+ * BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
+ * USE OF THIS SOFTWARE, EVEN IF NOT ADVISED OF THE POSSIBILITY OF SUCH
+ * DAMAGE.
+ */
+
+#ifndef X25519_H
+#define X25519_H
+
+#include "ext_headers.h"
+#include "lc_kyber.h"
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+int lc_x25519_keypair(struct lc_x25519_pk *pk, struct lc_x25519_sk *sk,
+		      struct lc_rng_ctx *rng_ctx);
+int lc_x25519_ss(struct lc_x25519_ss *ss, const struct lc_x25519_pk *pk,
+		 const struct lc_x25519_sk *sk);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif /* X25519_H */