+++ CHANGES.md
@@ -10,6 +10,8 @@ Changes 0.8.0-prerelease:
 
 * reduce memory footprint of Dilithium and Kyber
 
+* enhancement: Add Kyber-X25519 KEM
+
 Changes 0.7.0:
 
 * enhancement: add XDRBG256 - the SHAKE256-based DRNG discussed for SP800-90A